librbd/cache: Establish the framework to integrate RWL and SSD
- Create WriteLogCache class
- Rename ReplicatedWriteLog files to AbstractWriteLog and
  modify the I/O method names
- fix the test
- Modify CMakeLists.txt to add newly created classes
