+++ src/librbd/cache/rwl/InitRequest.cc
@@ -9,7 +9,7 @@
 
 #if defined(WITH_RBD_RWL)
 #include "librbd/cache/rwl/ImageCacheState.h"
-#include "librbd/cache/ReplicatedWriteLog.h"
+#include "librbd/cache/WriteLogCache.h"
 #endif // WITH_RBD_RWL
 
 #include "librbd/cache/Utils.h"
@@ -76,8 +76,8 @@ void InitRequest<I>::get_image_cache_state() {
   switch(cache_type) {
     case cache::IMAGE_CACHE_TYPE_RWL:
       m_image_ctx.image_cache =
-        new librbd::cache::ReplicatedWriteLog<I>(m_image_ctx,
-                                                 cache_state);
+        new librbd::cache::WriteLogCache<I>(m_image_ctx,
+                                            cache_state);
       break;
     default:
       delete cache_state;
